configured retrofit

+++ app/src/main/java/com/arifur/weatherappusingkotlin/model/Daily.kt
@@ -75,6 +75,6 @@ class Daily {
 
     @SerializedName("uvi")
     @Expose
-    var uvi: Int? = null
+    var uvi: Double? = null
 
 }
